Improved Pieces abstraction
