+++ core/development/logic/items/Item.test.ts
@@ -1,5 +1,4 @@
 import { HEAVEN_BOARD_ID, OVERWORLD_BOARD_ID } from '../../Constants';
-import { Position, Square } from '../pieces/PiecesUtilities';
 import { PiggyBank } from './PiggyBank';
 import { Player } from '../players/Player';
 import { onPlayerAction } from '../PieceLogic';
@@ -8,6 +7,8 @@ import { Rook } from '../pieces/Rook';
 import { Trap } from './Trap';
 import { PlayerColor } from '../players/types/PlayerColor';
 import { PlayerInventory } from '../inventory/PlayerInventory';
+import { Position } from '../pieces/types/Position';
+import { Square } from '../pieces/types/Square';
 
 const whitePlayer = new Player(PlayerColor.WHITE, new PlayerInventory());
