+++ core/development/logic/items/Trap.ts
@@ -2,10 +2,10 @@ import { BaseItem } from './abstract/Item';
 import { trapResource } from '../../ui/Resources';
 import { spawnItemOnBoard } from '../../LogicAdapter';
 import { game } from '../../Game';
-import { Position } from '../pieces/PiecesUtilities';
 import { Log, MovementLog } from '../../ui/logs/Log';
 import { Logger } from '../../ui/logs/Logger';
 import { ItemActionResult } from './types/ItemActionResult';
+import { Position } from '../pieces/types/Position';
 
 export class Trap extends BaseItem {
   constructor(position?: Position) {
