+++ core/development/logic/pieces/Bishop.ts
@@ -1,9 +1,9 @@
 import { bishopResource } from '../../ui/Resources';
 import { Player } from '../players/Player';
-import { Position } from './PiecesUtilities';
 import { getPieceByPosition } from '../Utilities';
 import { PlayerColor } from '../players/types/PlayerColor';
 import { BasePiece } from './abstract/BasePiece';
+import { Position } from './types/Position';
 
 export class Bishop extends BasePiece {
   constructor(player: Player, position?: Position) {
