+++ core/development/logic/pieces/Golem.test.ts
@@ -2,11 +2,11 @@ import { game } from '../../Game';
 import { HEAVEN_BOARD_ID, OVERWORLD_BOARD_ID } from '../../Constants';
 import { onPlayerAction } from '../PieceLogic';
 import { Player } from '../players/Player';
-import { Position } from './PiecesUtilities';
 import { Pawn } from './Pawn';
 import { Golem } from './Golem';
 import { PlayerColor } from '../players/types/PlayerColor';
 import { PlayerInventory } from '../inventory/PlayerInventory';
+import { Position } from './types/Position';
 
 const whitePlayer = new Player(PlayerColor.WHITE, new PlayerInventory());
 const blackPlayer = new Player(PlayerColor.BLACK, new PlayerInventory());
