+++ core/development/logic/pieces/Rook.test.ts
@@ -7,11 +7,11 @@ import {
 } from '../../Constants';
 import { onPlayerAction } from '../PieceLogic';
 import { Player } from '../players/Player';
-import { Position } from './PiecesUtilities';
 import { Rook } from './Rook';
 import { Pawn } from './Pawn';
 import { PlayerColor } from '../players/types/PlayerColor';
 import { PlayerInventory } from '../inventory/PlayerInventory';
+import { Position } from './types/Position';
 
 const whitePlayer = new Player(PlayerColor.WHITE, new PlayerInventory());
 const blackPlayer = new Player(PlayerColor.BLACK, new PlayerInventory());
