+++ core/development/logic/pieces/Rook.ts
@@ -1,10 +1,10 @@
 import { rookResource } from '../../ui/Resources';
 import { Player } from '../players/Player';
-import { Position } from './PiecesUtilities';
 import { getPieceByPosition } from '../Utilities';
 import { Golem } from './Golem';
 import { PlayerColor } from '../players/types/PlayerColor';
 import { BasePiece } from './abstract/BasePiece';
+import { Position } from './types/Position';
 
 export class Rook extends BasePiece {
   constructor(player: Player, position?: Position) {
