+++ core/development/logic/pieces/types/Position.ts
@@ -1,0 +1,4 @@
+export type Position = {
+  coordinates: [number, number];
+  boardId: string;
+};
