+++ core/development/logic/pieces/types/Square.ts
@@ -1,0 +1,7 @@
+import { BasePiece } from '../abstract/BasePiece';
+import { Position } from './Position';
+
+export type Square = {
+  position: Position;
+  occupent?: BasePiece;
+};
